+++ pkgs/test_core/lib/backend.dart
@@ -1,0 +1,11 @@
+// Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+export 'package:test_api/src/backend/metadata.dart' show Metadata;
+export 'package:test_api/src/backend/platform_selector.dart'
+    show PlatformSelector;
+export 'package:test_api/src/backend/runtime.dart' show Runtime;
+export 'package:test_api/src/backend/suite_platform.dart' show SuitePlatform;
+
+export 'src/runner/parse_metadata.dart' show parseMetadata;